change swr computation to voltage instead of power since detectors are in square law range, fix more bugs and return to coarse/fine tune algorithm, speed up relay, swr measure

+++ Arduino_tuner/Arduino_tuner/prototypes.h
@@ -27,7 +27,7 @@ extern relayState cIn;
 extern char control_port_buffer[COMMAND_BUFFER_SIZE];
 extern int control_port_buffer_index;
 extern unsigned long last_serial_receive_time;
-extern int fwd, ref;
+extern float fwd, ref;
 
 void ToggleRelay(relayState* CurrentState, int toggle = 1);
 void setStates(relayState* stateArray, byte newValue);
